requirements.txt
streamlit>=1.31.0
playwright>=1.40.0
python-dotenv>=1.0.0
pandas>=2.1.0
google-genai>=0.2.0
